requirements.txt
streamlit
matplotlib
pandas
pandas
PyGithub

